auto-monitor/requirements.txt
fastapi
uvicorn
httpx
apscheduler
python-dotenv
pydantic-settings